overlay: persist settings
So there's a bit more at play here. The booleans for showing the windows
have been moved to a new OverlayOptions struct (again), which will now
contain all settings that are meant to be persisted across application
restarts.

*State* structs will now be volatile, and only used for maintaining state
within the use of the application, but not persisted. This will mostly
be needed for things like dive/combat/player selection, which cannot be
loaded at runtime as the log may not contain the same information at
next launch.

Signed-off-by: Eric Richter <richteer@lastprime.net>