refactor(parser): use an Enum for handling Aspects
Resisted this, but now using an enum internally for aspects, as there
are enough bits of information to be attached to them that carrying a
bunch of patchwork functions would become too unwieldy.

Define an enum Aspect, which implements Display for getting the
human-friendly string, and a function .from_id() to parse the "C00"
string into the aspect.

Use an Unknown(String) aspect so that theoretically future aspects can
still be parsed and handled somewhat correctly, while not requiring them
to be hardcoded in. This should largely just reduce the number of
string comparisons needed in rendering code.

Signed-off-by: Eric Richter <richteer@lastprime.net>

Signed-off-by: Eric Richter <richteer@lastprime.net>