feat(overlay): add support for custom aspect colors
Allow custom colors to be set via a new window opened via settings.
Custom aspect colors are persisted in the same struct as options, and are
only stored if overridden from default.

Helper interfaces have been implemented on various structs to make using
the default-or-custom colors more ergonomic, but this implementation
really could use some cleanup.

Furthermore, it is not currently possible to set colors to override
currently unknown Aspects, this should be implemented in the future.

Signed-off-by: Eric Richter <richteer@lastprime.net>