perf(main): put backlog parsing in its own thread to not slow full application load
The UI takes over the main thread, so it has to be initialized last.
Therefore, any program initialization should be as quick as possible, or
done in the background after the UI loads.

This should probably be followed up with a backlog parsing indicator in
the UI, to let users know that the data is still loading.

Signed-off-by: Eric Richter <richteer@lastprime.net>