refactor(overlay)!: window overhaul part 1: move to using a trait-based display method
Instead of calling a function to render a window, instead use a struct
that defines how a particular window is rendered. That is, the "History"
window is managed by the "HistoryWindow" struct, that implements the
"WindowDisplay" trait.

This way, all (dynamic, more on that later) windows can be rendered by a
simple loop over a vector containing all the declared windows. For now,
this will be the same set of windows as the older version.

Furthermore, all windows now contain their own state, and no longer need
to store their state in a centralized configuration struct. This is
mostly intended to allow for multiple windows of the same type, but with
different configurations. For example, two individual combat (skills)
windows for two different players to be handled at the same time.

Signed-off-by: Eric Richter <richteer@lastprime.net>