refactor(overlay)!: window overhaul part 2: re-enable persisting windows
BREAKING: Users must click "Reset Default Windows" to restore the
previously default windows, as they have not been defined in existing
stored state.

Since changing to a more dynamic window scheme, persistence was going to
be a bit trickier as each window is handled with its own unique type.
This commit re-introduces the ability to store and load window states by
persisting the internal windows and enabled_windows lists in the same
manner as the overall Overlay settings.

This means two things: options set in windows will now be persisted (like
the bar settings in History), and also the set of windows that are
defined and displayed will also be shown.

As the breaking warning above also states, this also means that for users
that have already launched the application once before and have generated
options, they will now have NO windows defined or enabled, as those lists
were not yet generated. A "Reset Default Windows" button has been
included for convenience.

Signed-off-by: Eric Richter <richteer@lastprime.net>