feat(overlay)!: add support for creating (and deleting) multiple windows of the same type
BREAKING: All windows should probably be recreated with the reset button
to ensure the IDs are set correctly.

Windows have now been overhauled (again) to be contained in an
OverlayWindow struct. This mostly just smooths out the rendering magic,
since it now manages the ID generation if creating new, or stores it when
serialized. As such, the ID method has been removed from the
WindowDisplay trait, as it is no longer needed to be, and should not be,
implemented by a window.

Since creation of multiple windows may lead to duplicate windows, the
window list in settings now displays the windows by name, and the name
may change in the event of changing window settings. For example, the
Individual windows now show the name of the player if not using the
default.

Hovering over the window in the settings list will highlight the window
it refers to in white if over the checkbox, and in red if over the
delete button.

There's probably some other small refactors in here, but most likely not
really worth mentioning.

Signed-off-by: Eric Richter <richteer@lastprime.net>