refactor(overlay)!: merge individual combat and dive windows
For a long time, the split between dive and combat logic handling has
been a nuisance as most of the behavior after selecting the data is the
same.

As part of the big refactor effort, merge the individual skills window
into a single window that switches based on an internal mode. This now
reduces the required code to handle the two dramatically. It may be a
consideration to de-merge them back if for some reason it's better to
have separate types for each, but that would most likely be done as a
higher level of indirection on top of the existing merge.

Also the default player is now the POV character, if one is set.

Signed-off-by: Eric Richter <richteer@lastprime.net>