refactor(overlay)!: factor out dive/combat selection logic
Windows will often need to handle different behavior between dives and
combats, so factor out the selectable switch between them to make it
slightly easier.

Also, implement for free the ability to create a new window with the mode
already selected -- mostly for the default set of windows where one-each
is preferred.

Signed-off-by: Eric Richter <richteer@lastprime.net>