feat: add -w/--window argument to overlay in a window
In lieu of solving the monitor selection problem, or implementing a
proper windowed mode, add an argument that renders the same overlay in
an opaque window, so that it can at least be probably used on
odd setups, or those with multiple monitors.

Also reconfigures the default settings, the original settings for debug
mode now ARE WindowedOverlay mode.

Signed-off-by: Eric Richter <richteer@lastprime.net>