refactor(overlay/extractors): add trait for stat selection, so that the options and behaviors can be reused
Like the dive/combat split, add a trait that windows can implement to
provide them with a stat combat selection box helper, as well as the
`.extract_stat` function, that handles extracting the configured stat
from a PlayerStats struct.

This has two purposes:
 1. allow the extraction logic to maintain its own state within foreign
    structs
 2. allow reuse of this logic for other windows.

Now that the extraction logic maintains its own state, it can also
allow extractors to have additional parameters. Most notably, this opens
the way for "Status Effects Applied" to conditionally show its own
selection box for which status effect to extract.

This also allows the History window to inherit all of the work done for
the soon-to-be-renamed "Group Damage" for minimal rework.

Signed-off-by: Eric Richter <richteer@lastprime.net>