refactor(overlay/extractor): merge status effects into one extractor, add extra status effect selection option
Since the extractor for the status effects are all basically the same,
merge that logic into one extractor that takes in a parameter to separate
them. Add a selection box for that additional parameter, so the user can
select which status effect to display.

This now doesn't require the status effect extractors to be hardcoded,
although the current allowed set of status effects still are. Future work
will need to be done to make this filter configurable (there are a LOT
of status effects, so showing all will be annoying).

Furthermore, implement the Display trait on StatSelectionState, so
windows that use the current state as a title can show more specific
information when extractors have parameters.

Signed-off-by: Eric Richter <richteer@lastprime.net>