feat(overlay/history): implement stat selection for history
Like in the still-needs-to-be-renamed Group Damage window, the history
window now has the option to select what stat you want compared across
combats in a dive.

This was mostly straightforward, though required the helper functions to
be promoted to methods on the history window, so that they had access to
the extraction function.

Signed-off-by: Eric Richter <richteer@lastprime.net>