style(history): use selectable values instead of combobox for mode, re-enable totals in percent mode
To break up the mess of ComboBoxes in history, use a row of selectable
values instead for quick switching of display mode.

Since stacked/percent mode are basically the same, further unify them by
using the same "Show Totals" option in percent, by forcing the y value to
always be 100.0 in percent mode.

Signed-off-by: Eric Richter <richteer@lastprime.net>