refactor(main): move all log reading -> parsing logic into its own crate
It has been a long pain point to have the main function so cluttered with
initialization of threads, channels, etc.

Rewrite the LogReader struct to hold ALL state regarding the reading of
the log file, the file watcher, the parser, and the datalog.

This now also opens up the LogReader to be more easily accessible to the
overlay, so that it may be queried, restarted, configured, etc.

Signed-off-by: Eric Richter <richteer@lastprime.net>