feat(overlay/group): add option for custom format strings to group stats
Now that windows can be dynamically made, and can dynamically set what
items to display, the text on plot bars should also be able to be
dynamic.

This is currently implemented in a similar fashion as some other
per-window options, but it might be a bit over the top and unnecessary.

Two dependent crates have also been pulled in, it might be worth
re-evaluating how important they actually are.

Signed-off-by: Eric Richter <richteer@lastprime.net>