feat(overlay/skilltotals): add option for custom format strings to skill totals
Like group stats, implement the same custom formatting logic to skill
total windows. This might be the deathknell for the original intent of
having a trait handle the options implementation for the window
implementor as now a new option needs to be handled. Either that all
needs to be implemented as some kind of macro, or the concept should be
abandoned.

Signed-off-by: Eric Richter <richteer@lastprime.net>