feat(overlay/skilltotals): add option to merge upgraded skills
When ascending a binding, it is picked up by the parser as a separate
skill. This effectively means that the base binding will never increase
its damage count, and with the power scaling in a dive, also means that
it will likely be outpaced quickly by the ascended variant.

Since this may or may not be useful information, add an option that
attempts to merge the upgraded and non-upgraded skills together.

Signed-off-by: Eric Richter <richteer@lastprime.net>