test(overlay): add test case for catching config breakage
In the future, user's configuration should avoid resetting automatically
if at all possible. Add a test case that checks for the inability to
parse the storage configuration, which would result in reverting to
default settings.

This should be updated whenever there is a new feature added that may
cause this to break.

Signed-off-by: Eric Richter <richteer@lastprime.net>